+++ RELEASE_NOTES.md
@@ -1,5 +1,6 @@
 # Unreleased
   - *SPI change* `@Json String` database type mappers now use `@EncodedJson String` instead (#1953)
+  - Sql4JSqlLogger: fix NPE when using Script
 
 # 3.24.1
   - fix Bean property arguments type being erased on generic beans